% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_helper.R
\name{abundance_estimates}
\alias{abundance_estimates}
\title{Calculate abundance estimates}
\usage{
abundance_estimates(obj, abund, type = c("BroodYear", "ReleaseYear", "None"))
}
\arguments{
\item{obj}{a capture history data frame (see documentation for s4t_ch)}

\item{abund}{a vector of sites that identify kelts (i.e. adult fish ladders)}

\item{type}{the type of summarization for the abundance estimates. BroodYear
summarizes by broodyear (s - a1), ReleaseYear summarizes by time (s),
and None does not summarize.}
}
\value{
a capture history data frame without observations following
observations at the specified sites.
}
\description{
Calculate abundance estimates of cohorts
}
