% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_summary.R
\name{AIC.s4t_ageclass_model}
\alias{AIC.s4t_ageclass_model}
\title{Compute and return AIC of fitted model objects}
\usage{
\method{AIC}{s4t_ageclass_model}(object, ..., k = 2)
}
\arguments{
\item{object}{A \code{s4t_ageclass_model} model object}

\item{...}{Optionally more fitted model objects}

\item{k, }{The penalty parameter, taken to be 2. Not used but needed for
generic consistency}
}
\value{
If one object is provided, just returns a numeric value with the
corresponding AIC. If more than one is provided, it returns a \code{data.frame}
with rows corresponding to the objects and columns representing
the number of parameters estimated (\code{df}), and the AIC
}
\description{
Compute AIC of one or more fitted \code{s4t_ageclass_model} model objects
}
\examples{
\dontrun{
sim.dat <- sim_simple_s4t_ch(N = 2000)
m1 <- fit_ageclass(ageclass_formula = ~ FL + obs_time,
                   s4t_ch = sim.dat$s4t_ch)
m2 <- fit_ageclass(ageclass_formula = ~ FL + I(FL^2) + obs_time,
                   s4t_ch = sim.dat$s4t_ch)
AIC(m1, m2)
}

}
