% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_summary.R
\name{AIC.s4t_cjs}
\alias{AIC.s4t_cjs}
\title{Compute and return AIC of fitted model objects}
\usage{
\method{AIC}{s4t_cjs}(object, ..., k = 2)
}
\arguments{
\item{object}{A \code{s4t_cjs} model object}

\item{...}{Optionally more fitted model objects}

\item{k, }{The penalty parameter, taken to be 2. Not used but needed for
generic consistency}
}
\value{
If one object is provided, just returns a numeric value with the
corresponding AIC. If more than one is provided, it returns a \code{data.frame}
with rows corresponding to the objects and columns representing
the number of parameters estimated (\code{df}), and the AIC
}
\description{
Compute AIC of one or more fitted \code{s4t_cjs} model objects
}
