% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_summary.R
\name{anova.s4t_ageclass_model}
\alias{anova.s4t_ageclass_model}
\title{AIC table for ageclass model fits}
\usage{
\method{anova}{s4t_ageclass_model}(object, ...)
}
\arguments{
\item{object}{object of class \code{s4t_ageclass_model}}

\item{...}{object of class \code{s4t_ageclass_model}}
}
\value{
An object of class \code{anova} inheriting from class \code{data.frame} ???
}
\description{
Compute analysis of variance table.
}
