% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_summary.R
\name{anova.s4t_cjs_ml}
\alias{anova.s4t_cjs_ml}
\title{ANOVA table for mark-recapture model fits}
\usage{
\method{anova}{s4t_cjs_ml}(object, ...)
}
\arguments{
\item{object}{object of class \code{s4t_cjs_ml}}

\item{...}{object of class \code{s4t_cjs_ml}}
}
\value{
An object of class \code{anova} inheriting form class \code{data.frame} ???
}
\description{
Compute analysis of variance table.
}
