% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ch.R
\name{clean_s4t_ch_obs}
\alias{clean_s4t_ch_obs}
\title{Clean the capture history of a space4time capture history object}
\usage{
clean_s4t_ch_obs(s4t_ch)
}
\arguments{
\item{s4t_ch}{a \code{s4t_ch} object.}
}
\value{
a \code{clean_s4t_ch} object.
}
\description{
A short description...
}
\details{
Additional details...
}
