% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.s4t_ageclass_model}
\alias{coef.s4t_ageclass_model}
\alias{coefficients.s4t_ageclass_model}
\title{Extract fitted model coefficients}
\usage{
\method{coef}{s4t_ageclass_model}(object, ...)

\method{coefficients}{s4t_ageclass_model}(object, ...)
}
\arguments{
\item{object}{a fitted model object from \code{s4t_ageclass_model()}}

\item{...}{Other arguments. Not used (needed for generic consistency)}
}
\value{
A named vector of coefficients
}
\description{
\code{coef} extracts fitted model coefficients from
fitted model objects. \code{coefficients.s4t_ageclass_model} is an alias for it
}
\examples{
# not run

}
