% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.s4t_cjs}
\alias{confint.s4t_cjs}
\title{Confidence intervals for fitted model parameters for s4t_cjs}
\usage{
\method{confint}{s4t_cjs}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{s4t_cjs}}

\item{parm}{A specification of what parameters are to be
given confidence intervals. A character vector or names.}

\item{level}{The confidence level. Default is \code{0.95}}

\item{...}{additional args. Not used}
}
\value{
Gaussian-based confidence intervals for the fixed
effect coffiecients based on the confidence level
specified by \code{level}. Confidence intervals are on the
logit scale.
}
\description{
Computes Wald confidence intervals for one or more parameters in a fitted
model object
}
