% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_loo.R
\name{extract_log_lik_s4t}
\alias{extract_log_lik_s4t}
\title{Extract pointwise log-likelihood from a fitted s4t_cjs_rstan model}
\usage{
extract_log_lik_s4t(mod)
}
\arguments{
\item{mod}{a \code{s4t_cjs_rstan} object}
}
\value{
an \code{S} by \code{N} matrix of post-warmup extracted draws, where \code{S} is
the size of the posterior sample, and \code{N} is the number of unique observations.
}
\description{
Function for extracting and de-marginalizing the pointwise log-likelihood matrix
from a s4t_cjs_rstan object. See loo::extract_log_lik.
}
