% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ageclass.R
\name{fit_ageclass}
\alias{fit_ageclass}
\title{Fit ageclass model using ordinal regression}
\usage{
fit_ageclass(age_formula = ~1, s4t_ch)
}
\arguments{
\item{age_formula}{a \code{formula} object}

\item{s4t_ch}{a \code{s4t_ch} object}
}
\value{
a fitted \code{s4t_ageclass_model} object
}
\description{
add description
}
