% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cjs.R
\name{fit_s4t_cjs_ml}
\alias{fit_s4t_cjs_ml}
\title{Fit space-for-time mark-recapture model in likelihood framework}
\usage{
fit_s4t_cjs_ml(
  p_formula,
  theta_formula,
  ageclass_formula,
  cov_p = NULL,
  cov_theta = NULL,
  groups = NULL,
  s4t_ch,
  ndeps = 0.001,
  lmm = 5,
  maxit = 500,
  fixed_age = FALSE
)
}
\arguments{
\item{p_formula}{a formula for detection probabilities using default indices
(a1, a2, s, t, j, k, r, and g) or covariates.}

\item{theta_formula}{a formula for transition probabilities using default indices
(a1, a2, s, t, j, k, r, and g) or covariates.}

\item{ageclass_formula}{a formula for the effect structure of the age-class
sub-model.}

\item{cov_p}{a \code{data.frame} or \code{list} of \code{data.frame}'s containing the covariates
for p \verb{a1,a2,j,k,s,t,r,g} indices. See details.}

\item{cov_theta}{a \code{data.frame} or \code{list} of \code{data.frame}'s containing the covariates
for theta \verb{a1,a2,j,k,s,t,r,g} indices. See details.#' @param groups a \code{character} vector containing the names of the covariates that comprise the groups.
Default is no groups (\code{groups = NULL}).}

\item{groups}{a \code{character} vector containing the names of the covariates that comprise the groups.}

\item{s4t_ch}{a \code{s4t_ch} object}

\item{ndeps}{a \code{numeric} value passed to \code{optim}. See \code{optim} documentation.}

\item{lmm}{an \code{integer} value passed to \code{optim}. See \code{optim} documentation.}

\item{maxit}{an \code{integer} value passed to optim. See \code{optim} documentation.}

\item{fixed_age}{a \code{logical} value that determines whether the age-class model will be run
as a separate model (\code{TRUE}) or whether it is estimated along with the mark-recapture model (\code{FALSE}).}
}
\value{
a \code{s4t_cjs_ml} object.
}
\description{
Fits the time- and age-specific space-for-time mark-recapture
model using maximum likelihood.
}
\details{
Additional details...
}
\examples{
\dontrun{
sim.dat <- sim_simple_s4t_ch(N = 2000)
m1 <- fit_s4t_cjs_ml(p_formula = ~ t,
                     theta_formula = ~ a1 * a2 * s * j,
                     ageclass_formula = ~ FL,
                     fixed_age = TRUE,
                     s4t_ch = sim.dat$s4t_ch)
}



}
