% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cjs.R
\name{fit_s4t_cjs_rstan}
\alias{fit_s4t_cjs_rstan}
\title{Fit space-for-time mark-recapture model in Bayesian framework}
\usage{
fit_s4t_cjs_rstan(
  p_formula,
  theta_formula,
  ageclass_formula,
  cov_p = NULL,
  cov_theta = NULL,
  groups = NULL,
  s4t_ch,
  chains = 3,
  warmup = 500,
  iter = 1000,
  fixed_age = TRUE,
  ...
)
}
\arguments{
\item{p_formula}{a formula for detection probabilities using default indices
(a1, a2, s, t, j, k, r, and g) or covariates.}

\item{theta_formula}{a formula for transition probabilities using default indices
(a1, a2, s, t, j, k, r, and g) or covariates.}

\item{ageclass_formula}{a formula for the effect structure of the age-class
sub-model.}

\item{cov_p}{a \code{data.frame} or \code{list} of \code{data.frame}'s containing the covariates
for p \verb{a1,a2,j,k,s,t,r,g} indices. See details.}

\item{cov_theta}{a \code{data.frame} or \code{list} of \code{data.frame}'s containing the covariates
for theta \verb{a1,a2,j,k,s,t,r,g} indices. See details.}

\item{groups}{a \code{character} vector containing the names of the covariates that comprise the groups.}

\item{s4t_ch}{a \code{s4t_ch} object}

\item{chains}{an \code{integer} of the number of chains to run.}

\item{warmup}{an \code{integer} of the number of warmup iterations.}

\item{iter}{an \code{integer} of the number of total (warmup + actual) iterations.}

\item{fixed_age}{a \code{logical} object that determines whether the ageclass model will be run
as a separate model (TRUE) or whether it is estimated along with the CJS model (FALSE).}

\item{...}{further arguents to pass to \code{rstan::sampling}}
}
\value{
a \code{s4t_cjs_rstan} object.
}
\description{
Fits the time- and age-specific space-for-time mark-recapture
model in a Bayesian framework using Stan.
}
\examples{
\dontrun{
sim.dat <- sim_simple_s4t_ch(N = 2000)
m1 <- fit_s4t_cjs_rstan(p_formula = ~ t,
                        theta_formula = ~ a1 * a2 * s * j,
                        ageclass_formula = ageclass ~ FL,
                        fixed_age = TRUE,
                        s4t_ch = sim.dat$s4t_ch)
}

}
