% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_loo.R
\name{loo.s4t_cjs_rstan}
\alias{loo.s4t_cjs_rstan}
\title{Efficient approximate leave-one-out cross-validation (LOO)}
\usage{
\method{loo}{s4t_cjs_rstan}(x, ...)
}
\arguments{
\item{x}{a \code{s4t_cjs_rstan} object}

\item{...}{passed to loo::loo}
}
\value{
a \verb{ADD DETAILS} object
}
\description{
Efficient approximate leave-one-out cross-validation (LOO)
}
