% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ageclass.R
\name{plot.s4t_ageclass_model}
\alias{plot.s4t_ageclass_model}
\title{Plot confusion matrix for fit_ageclass fits}
\usage{
\method{plot}{s4t_ageclass_model}(x, param, ...)
}
\arguments{
\item{x}{a \code{s4t_ageclass_model} object}

\item{param}{not used.}

\item{...}{Not used.}
}
\value{
ggplot2 style figure
}
\description{
add description. Currently only does confusion matrix.
}
