% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCH}
\alias{plotCH}
\title{Plot summary of s4t_cjs_ch object}
\usage{
plotCH(x, ...)
}
\arguments{
\item{x}{a \code{s4t_cjs_ch} object}

\item{...}{not used}
}
\value{
a ggplot2 type figure of the observed transition
}
\description{
add description
}
