% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCov.R
\name{plotCov}
\alias{plotCov}
\title{Effect plot for s4t_cjs_rstan object}
\usage{
plotCov(mod, param, ...)
}
\arguments{
\item{mod}{a \code{s4t_cjs_rstan} object}

\item{param}{a \code{character} of the parameter}

\item{...}{not used at the moment}
}
\value{
a ggplot2 type figure of the transition probabilities
}
\description{
add description
}
