% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSurvival}
\alias{plotSurvival}
\title{Plot s4t_cjs or s4t_cjs_rstan object results}
\usage{
plotSurvival(x, textsize = 3, ...)
}
\arguments{
\item{x}{a \code{s4t_cjs} or \code{s4t_cjs_rstan} object}

\item{textsize}{an integer for the font size to pass to \code{geomtextpath::geom_textsegment()}}

\item{...}{not used}
}
\value{
a ggplot2 type figure of the apparent survivals
}
\description{
add description
}
