% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTransitions}
\alias{plotTransitions}
\title{Plot transition probability estimates from fitted \code{s4t_cjs_ml} or \code{s4t_cjs_rstan}objects}
\usage{
plotTransitions(x, textsize = 3, ...)
}
\arguments{
\item{x}{a \code{s4t_cjs} or \code{s4t_cjs_rstan} object}

\item{textsize}{an integer for the font size to pass to \code{geomtextpath::geom_textsegment()}}

\item{...}{passed to \code{dplyr::filter()} for selecting particular survivals. Filters
apparent survivals from \code{s4t_cjs_ml$apparent_surv} or \code{s4t_cjs_rstan$apparent_surv}}
}
\value{
a \code{ggplot2} figure showing estimated transition probabilities.
}
\description{
Plot transition probability estimates (which include movement and survival
probabilities) from fitted \code{s4t_cjs_ml} or \code{s4t_cjs_rstan}objects.
If the package \code{geomtextpath} is installed, the figure is more visually appealing.
}
\examples{
\dontrun{
 sim.dat <- sim_simple_s4t_ch(N = 2000)
 m1 <- fit_s4t_cjs_rstan(p_formula = ~ t,
                         theta_formula = ~ a1 * a2 * s * j,
                         ageclass_formula = ageclass ~ FL,
                         fixed_age = TRUE,
                         s4t_ch = sim.dat$s4t_ch)
plotTransitions(m1)
}

}
