% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.s4t_ageclass_model}
\alias{predict.s4t_ageclass_model}
\title{Predict method for fit_ageclass fits}
\usage{
\method{predict}{s4t_ageclass_model}(object, newdata, type = c("prob", "class", "cum.prob"))
}
\arguments{
\item{object}{a \code{s4t_ageclass_model} object}

\item{newdata}{add description.}

\item{type}{a \code{character} of one of the three option \code{prob}, \code{class},
or \code{cum.prob}}
}
\value{
depends on the type.
}
\description{
add description
}
