% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cov_s4t_ch}
\alias{print.cov_s4t_ch}
\title{Print summary of \code{cov_s4t_ch} object}
\usage{
\method{print}{cov_s4t_ch}(x, ...)
}
\arguments{
\item{x}{\code{s4t_ch} object}

\item{...}{passed to \code{print.tibble()}}
}
\description{
Returns the indices
}
