% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.s4t_ch}
\alias{print.s4t_ch}
\title{Print summary of \code{s4t_ch} object}
\usage{
\method{print}{s4t_ch}(x, ...)
}
\arguments{
\item{x}{\code{s4t_ch} object}

\item{...}{passed to \code{print()}}
}
\description{
Print summary of \code{s4t_ch} object.
}
