% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.s4t_cjs_ml}
\alias{print.s4t_cjs_ml}
\title{print s4t_cjs_ml}
\usage{
\method{print}{s4t_cjs_ml}(x, ...)
}
\arguments{
\item{x}{s4t_cjs_ml object}

\item{...}{for generic consistency}
}
\description{
print s4t_cjs_ml
}
