% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.s4t_cjs_rstan}
\alias{print.s4t_cjs_rstan}
\title{Print summary of \code{s4t_cjs_rstan}}
\usage{
\method{print}{s4t_cjs_rstan}(
  x,
  pars = NULL,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  digits_summary = 2,
  include = TRUE,
  ...
)
}
\arguments{
\item{x}{s4t_cjs_rstan object}

\item{pars}{vector of parameters to include (defaults to all parameters)}

\item{probs}{quantiles for credible intervals}

\item{digits_summary}{the number of significant digits to use when printing the summary}

\item{include}{logical scalar indicating whether to include the parameters named by the
pars argument.}

\item{...}{passed to \code{summary} method for s4t_cjs_rstan object.}
}
\description{
Print summary of \code{s4t_cjs_rstan}
}
