% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.s4t_config}
\alias{print.s4t_config}
\title{Print summary of \code{s4t_config} object}
\usage{
\method{print}{s4t_config}(x, ...)
}
\arguments{
\item{x}{\code{s4t_config} object}

\item{...}{passed to \code{print()}}
}
\description{
Print summary of \code{s4t_config} object.
}
