% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.s4t_cjs}
\alias{print.summary.s4t_cjs}
\title{print summary of s4t_cjs}
\usage{
\method{print}{summary.s4t_cjs}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{summary.s4t_cjs object}

\item{digits}{number of digits}

\item{...}{passed to}
}
\description{
print summary of s4t_cjs
}
