% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_helper.R
\name{read_DART_file}
\alias{read_DART_file}
\title{Read DART Basin TribPit observation files and combine with auxiliary age data}
\usage{
read_DART_file(
  filepath,
  aux_age_df,
  DART_config = "https://www.cbr.washington.edu/downloads/paramest/sites_config.txt",
  ...
)
}
\arguments{
\item{filepath}{the path to the DART Basin TribPit observation file}

\item{aux_age_df}{A data frame object in the same format as aux_age_df
(see documentation for \code{s4t_ch})}

\item{DART_config}{the path to a site configuration file \code{"site_config.txt"} generated by
Columbia Basin Research. If blank, defaults to \code{www.cbr.washington.edu/downloads/paramest/sites_config.txt}}

\item{...}{any additional file paths of DART files to include}
}
\value{
A list with two elements. The first is the formatted capture history data (\code{ch_df})
and the second is the age auxiliary data (\code{aux_age_df}).
}
\description{
Read DART Basin TribPit observation files and combine with auxiliary age data
}
\details{
Currently, only Dams in the Snake River are checked for transported fish.
}
