% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart_helper.R
\name{remove_kelt_obs}
\alias{remove_kelt_obs}
\title{Remove observations of kelts}
\usage{
remove_kelt_obs(ch_df, kelt_obssite)
}
\arguments{
\item{ch_df}{a capture history data frame (see documentation for s4t_ch)}

\item{kelt_obssite}{a vector of sites that identify kelts (i.e. adult fish ladders)}
}
\value{
a capture history data frame without observations following
observations at the specified sites.
}
\description{
Removes any observations following an observation at specified sites. Note that
the observations at the specified sites are retained.
}
