% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ch.R
\name{s4t_ch}
\alias{s4t_ch}
\title{Create capture history object}
\usage{
s4t_ch(ch_df, aux_age_df, s4t_config)
}
\arguments{
\item{ch_df}{\code{data.frame} object containing each capture. See LINK VIGNETTE and details.}

\item{aux_age_df}{\code{data.frame} containing auxiliary data for each individual. See LINK VIGNETTE and details.}

\item{s4t_config}{a \code{s4t_config} object created using \code{s4t_config()},
\code{linear_s4t_config}, or \code{simplebranch_s4t_config}.}
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{

ch_df <- data.frame(id = c(1,1,1,
                           2,2,
                           3,3,
                           4,
                           5,
                           6),
                     site = c("A","B","C",
                              "A","B",
                              "A","C",
                              "A",
                              "A",
                              "A"),
                     time = c(1,3,3,
                              2,3,
                              1,3,
                              2,
                              1,
                              1),
                     removed = c(FALSE,FALSE,FALSE,
                                 FALSE,FALSE,
                                 FALSE,FALSE,
                                 FALSE,
                                 FALSE,
                                 FALSE)
                      )

aux_age_df <- data.frame(id = 1:6,
                          obs_site = rep("A",6),
                          ageclass = c(1,2,1,1,2,1),
                          obs_time = c(1,2,1,2,1,1),
                          Covariate1 = c(3,1,2,1,2,1))

site_arr <- linear_s4t_config(sites_names = c("A","B","C"),
                              holdover_sites = c("A"),
                              min_a = c(1,1,1),
                              max_a = c(3,3,3))

ch <- s4t_ch(ch_df = ch_df,
             aux_age_df = aux_age_df,
             s4t_config = site_arr)




}
