% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ch.R
\name{s4t_cjs_ch}
\alias{s4t_cjs_ch}
\title{Create capture history object}
\usage{
s4t_cjs_ch(
  ch_df,
  aux_age_df,
  cov_df = NULL,
  min_a,
  max_a,
  sites_names,
  sites_config,
  holdover_config,
  sites_to_pool = NULL
)
}
\arguments{
\item{ch_df}{\code{data.frame} object containing each capture. See LINK VIGNETTE and details.}

\item{aux_age_df}{\code{data.frame} containing auxiliary data for each individual. See LINK VIGNETTE and details.}

\item{cov_df}{\code{NULL} or \code{data.frame} of covariates aligned with \verb{j,k,s,t,b,r} indices. See LINK VIGNETTE and details.}

\item{min_a}{a \code{vector} of the minimum ageclass individuals in a site can be.
Must be the same length and order as \code{sites_names}}

\item{max_a}{a \code{vector} of the maximum ageclass individuals in a site can be.
Must be the same length and order as \code{sites_names}}

\item{sites_names}{a character \code{vector} of the site names}

\item{sites_config}{a \code{matrix} that determine how sites are linked together.
Must be in the same order as sites_names. See LINK VIGNETTE and details.}

\item{holdover_config}{a \code{matrix} that determine whether individuals can holdover in between
sites. In the same format as sites_configs. See LINK VIGNETTE and details.}

\item{sites_to_pool}{a list of character vectors that contain the names of sites
to be pooled together and treated as one site. See LINK VIGNETTE and details.}
}
\description{
A short description...
}
\details{
Additional details...
}
\examples{

ch_df <- data.frame(id = c(1,1,1,
                           2,2,
                           3,3,
                           4,
                           5,
                           6),
                     site = c(1,2,3,
                              1,2,
                              1,3,
                              1,
                              1,
                              1),
                     time = c(1,3,3,
                              2,3,
                              1,3,
                              2,
                              1,
                              1),
                     removed = c(FALSE,FALSE,FALSE,
                                 FALSE,FALSE,
                                 FALSE,FALSE,
                                 FALSE,
                                 FALSE,
                                 FALSE)
                      )

aux_age_df <- data.frame(id = 1:6,
                          obs_site = rep(1,6),
                          ageclass = c(1,2,1,1,2,1),
                          obs_time = c(1,2,1,2,1,1),
                          Covariate1 = c(3,1,2,1,2,1))

ch <- s4t_cjs_ch(ch_df,aux_age_df,
                 min_a = c(1,1,1),
                 max_a = c(3,3,3),
                 sites_names = 1:3,
                 sites_config = matrix(c(0,1,0,
                                         0,0,1,
                                         0,0,0),
                                         nrow = 3,
                                         ncol = 3,
                                         byrow = TRUE,
                                         dimnames = list(1:3,1:3)),
                holdover_config = matrix(c(0,1,0,
                                           0,0,0,
                                           0,0,0),
                                           nrow = 3,
                                           ncol = 3,
                                           byrow = TRUE,
                                           dimnames = list(1:3,1:3)),
                )




}
