% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4t_config.R
\name{s4t_config}
\alias{s4t_config}
\title{Create space4time site configuration object (\code{s4t_config})}
\usage{
s4t_config(
  sites_names,
  sites_config,
  holdover_config,
  min_a,
  max_a,
  sites_to_pool = NULL
)
}
\arguments{
\item{sites_names}{a character \code{vector} of the site names}

\item{sites_config}{a \code{matrix} that determine how sites are linked together.
Must be in the same order as sites_names. See LINK VIGNETTE and details.}

\item{holdover_config}{a \code{matrix} that determine whether individuals can holdover in between
sites. In the same format as sites_configs. See LINK VIGNETTE and details.}

\item{min_a}{a \code{vector} of the minimum ageclass individuals in a site can be.
Must be the same length and order as \code{sites_names}}

\item{max_a}{a \code{vector} of the maximum ageclass individuals in a site can be.
Must be the same length and order as \code{sites_names}}

\item{sites_to_pool}{a named list of character vectors that contain the names of sites
to be pooled together and treated as one site. See examples, LINK VIGNETTE, and details.}
}
\value{
A \code{s4t_config} object, which contains
}
\description{
Create the space4time site configuration object (\code{s4t_config}) using
custom site configuration. For a simple configuration where
there is only one initial release site use \code{linear_s4t_config()},
which assumes that the sites are arranged in a simple sequence.
If there are multiple initial release sites that are immediately
connected to a simple sequence then \code{simplebranch_s4t_config()}
can be used. Otherwise, use this function to create a custom site
configuration. See details or (LINK VIGNETTE) for more detailed
instructions.
}
\details{
Additional details...
\code{sites_to_pool} is either a named list or a list of character vectors. If
it is unnamed, than the name, min_a, max_a values are used from the first
site listed. If the list is named, than the name, min_a, and max_a are
used from the site in the name.
}
\examples{
site_arrangement <- s4t_config(sites_names = c("A","B","C","D"),
                               min_a = c(1,1,1,1),
                               max_a = c(3,4,4,4),
                               sites_config = matrix(
                                       c(0,1,0,0,
                                         0,0,1,0,
                                         0,0,0,1,
                                         0,0,0,0),
                                         nrow = 4,
                                         ncol = 4,
                                         byrow = TRUE,
                                         dimnames = list(c("A","B","C","D"),
                                         c("A","B","C","D"))),
                               holdover_config = matrix(
                                        c(0,1,0,0,
                                          0,0,0,0,
                                          0,0,0,0,
                                          0,0,0,0),
                                          nrow = 4,
                                          ncol = 4,
                                          byrow = TRUE,
                                          dimnames = list(c("A","B","C","D"),
                                          c("A","B","C","D")))
)


}
