% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ch.R
\name{sim_simple_s4t_ch}
\alias{sim_simple_s4t_ch}
\title{Simulate simple capture history}
\usage{
sim_simple_s4t_ch(N = 500, max_obs_year = 2, prop_missing_age = 0.5)
}
\arguments{
\item{N}{An \code{integer} of the number of individuals}

\item{max_obs_year}{an \code{integer} of the number of total time intervals}

\item{prop_missing_age}{a \code{numeric} value between \code{0} and \code{1} of the proportion of
individuals that do not have known ages.}
}
\value{
a \code{list} containing capture history \code{data.frame}, \code{list} containing the
params.
}
\description{
Simulates a small capture history with 3 sites, where there can be holdovers
betweeen the first 1st and 2nd sites.
}
