% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4t_config.R
\name{simplebranch_s4t_config}
\alias{simplebranch_s4t_config}
\title{Create site configuration for simple branching site arrangement}
\usage{
simplebranch_s4t_config(
  sites_names,
  branch_sites,
  holdover_sites = NULL,
  min_a,
  max_a,
  sites_to_pool = NULL
)
}
\arguments{
\item{sites_names}{a character \code{vector} of the site names. Order indicates the
direction of movement.}

\item{branch_sites}{a character \code{vector} of the initial branching sites}

\item{holdover_sites}{a character \code{vector} of the sites where after passing,
individuals can holdover (take 1 or more time intervals) before reaching the next site.}

\item{min_a}{a \code{vector} of the minimum ageclass individuals in a site can be.
Must be the same length and order as \code{sites_names}}

\item{max_a}{a \code{vector} of the maximum ageclass individuals in a site can be.
Must be the same length and order as \code{sites_names}}

\item{sites_to_pool}{a named list of character vectors that contain the names of sites
to be pooled together and treated as one site. See LINK VIGNETTE and details.}
}
\description{
This is only for when there are two or more initial sites where individuals
can be released, and then the next site is for all sites. After this
shared site, there should be at least one additional site in order to estimate
transitions and apparent survival (because detection probability cannot
be estimated for the last site).
}
