% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_summary.R
\name{summary.s4t_ageclass_model}
\alias{summary.s4t_ageclass_model}
\title{summary of s4t_ageclass_cjs object}
\usage{
\method{summary}{s4t_ageclass_model}(object, ...)
}
\arguments{
\item{object}{A s4t_ageclass_cjs object}
}
\value{
estimated parameters table and AIC
}
\description{
summary of s4t_ageclass_cjs object
}
