% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_summary.R
\name{summary.s4t_cjs_rstan}
\alias{summary.s4t_cjs_rstan}
\title{summary of s4t_cjs_rstan object}
\usage{
\method{summary}{s4t_cjs_rstan}(object, pars = NULL, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{object}{A s4t_cjs_rstan object.}

\item{pars}{regular expressions}

\item{probs}{a numerical vector of quantiles of interest.}

\item{...}{Additional arguments to pass to \verb{summary,stanfit-method}.}
}
\value{
Returns a named list with ...
}
\description{
Calls \verb{summary,stanfit-method}.
}
