% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{traceplot.s4t_cjs_rstan}
\alias{traceplot.s4t_cjs_rstan}
\title{Create traceplots from \code{s4t_cjs_rstan} object}
\usage{
traceplot.s4t_cjs_rstan(object, pars = NULL, ...)
}
\arguments{
\item{object}{a \code{s4t_cjs_rstan} object}

\item{pars}{a \code{character} object that parses parameters using regular
expressions}

\item{...}{not used}
}
\value{
a ggplot2 object of traceplots
}
\description{
add description
}
