% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.s4t_ageclass_model}
\alias{vcov.s4t_ageclass_model}
\title{Calculate variance-covariance matrix for a fitted model object}
\usage{
\method{vcov}{s4t_ageclass_model}(object, ...)
}
\arguments{
\item{object}{a \code{s4t_ageclass_model} object fitted using \code{fit_ageclass()}.}

\item{...}{Other arguments. Not used (needed for generic consistency)}
}
\value{
The variance-covariance matrix of coefficients from a fitted
\code{s4t_ageclass_model} object
}
\description{
Calculate variance-covariance matrix for a \code{s4t_ageclass_model} model object.
}
\examples{
\dontrun{
sim.dat <- sim_simple_s4t_ch(N = 2000)
m1 <- fit_ageclass(ageclass_formula = ~ FL
                   s4t_ch = sim.dat$s4t_ch)

vcov(m1)
}

}
