% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.s4t_cjs}
\alias{vcov.s4t_cjs}
\title{returns variance-covariance matrix}
\usage{
\method{vcov}{s4t_cjs}(object, ...)
}
\arguments{
\item{object}{A \code{s4t_cjs} object}

\item{...}{Other arguments. Not used (for generic consistency)}
}
\value{
The variance-covariance matrix for the CJS portion of the model
}
\description{
returns variance-covariance matrix
}
