% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.s4t_cjs_ml}
\alias{vcov.s4t_cjs_ml}
\title{Calculate variance-covariance matrix for a fitted model object}
\usage{
\method{vcov}{s4t_cjs_ml}(object, ...)
}
\arguments{
\item{object}{a \code{s4t_cjs_ml} object fitted using \code{fit_s4t_cjs_ml()}.}

\item{...}{Other arguments. Not used (needed for generic consistency)}
}
\value{
The variance-covariance matrix of coefficients from a fitted
\code{s4t_cjs_ml} object
}
\description{
Calculate variance-covariance matrix for a \code{s4t_cjs_ml} model object.
}
\examples{
\dontrun{
sim.dat <- sim_simple_s4t_ch(N = 2000)
m1 <- fit_s4t_cjs_ml(p_formula = ~ t,
                     theta_formula = ~ a1 * a2 * s * j,
                     ageclass_formula = ~ FL,
                     fixed_age = TRUE,
                     s4t_ch = sim.dat$s4t_ch)

vcov(m1)
}

}
