% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_loo.R
\name{waic.s4t_cjs_rstan}
\alias{waic.s4t_cjs_rstan}
\title{Widely applicable information criterion (WAIC)}
\usage{
\method{waic}{s4t_cjs_rstan}(x, ...)
}
\arguments{
\item{x}{a \code{s4t_cjs_rstan} object}

\item{...}{not used}
}
\value{
A named list of class ....
}
\description{
add description
}
