% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_covariates.R
\name{add_covariates}
\alias{add_covariates}
\title{Add covariates to a space4time capture history object}
\usage{
add_covariates(cov_df, s4t_ch)
}
\arguments{
\item{cov_df}{a \code{data.frame} or \code{list} of \code{data.frame}'s containing the covariates
for theta and p \verb{a1,a2,j,k,s,t,r,g} indices. See details.}

\item{s4t_ch}{a \code{s4t_ch} object.}
}
\value{
a \code{s4t_ch} object with covariates added
}
\description{
Add covariates to a space4time capture history object. If the covariate has
missing levels from the indices (\verb{a1,a2,j,k,s,t,r,g}), an indicator
variable is added that is named \code{"OFF_"} appended to the covariate name.
}
\details{
To show how covariates are added or what levels are required for the indices,
use extract_covariates.
}
