% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_summary.R
\name{extract_ageclass_fit}
\alias{extract_ageclass_fit}
\title{Extract s4t_ageclass_model object from s4t_cjs_rstan or s4t_cjs_ml}
\usage{
extract_ageclass_fit(object)
}
\arguments{
\item{object}{a s4t_cjs_rstan or s4t_cjs_ml object}
}
\value{
a s4t_ageclass_model object
}
\description{
Add description
}
