% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_covariates.R
\name{extract_covariates}
\alias{extract_covariates}
\title{Extract indices and covariates from space4time capture history objects}
\usage{
extract_covariates(s4t_ch)
}
\arguments{
\item{s4t_ch}{a \code{s4t_ch} object.}
}
\value{
a list of class \code{cov_s4t_ch} that contains the indices for $theta$
and $p$ parameters.
}
\description{
Extract indices and covariates from space4time capture history objects. Note that
grouping indices are excluded.
}
