% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_covariates.R
\name{extract_covariates}
\alias{extract_covariates}
\title{Add covariates to a space4time capture history object}
\usage{
extract_covariates(s4t_ch)
}
\arguments{
\item{s4t_ch}{a \code{s4t_ch} object.}

\item{cov_df}{a \code{data.frame} or \code{list} of \code{data.frame}'s containing the covariates
for theta and p \verb{a1,a2,j,k,s,t,r,g} indices. See details.}
}
\value{
a \code{s4t_ch} object with covariates added
}
\description{
A short description...
}
\details{
Additional details...
}
