% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ageclass.R
\name{fit_ageclass}
\alias{fit_ageclass}
\title{Fit age-class model using ordinal regression}
\usage{
fit_ageclass(age_formula = ~1, s4t_ch)
}
\arguments{
\item{age_formula}{a one- or two-sided \code{formula} describing the effect
structure of the model. Note that obs_time is treated as factor}

\item{s4t_ch}{a \code{s4t_ch} object}
}
\value{
a fitted \code{s4t_ageclass_model} object that contains the estimated
parameters, AIC, negative log likelihood, and call.
}
\description{
Fit age-class ordinal regression model with logit link and
flexible threshold parameters.
}
\details{
To enforce obs_time to be treated as a continuous variable, the formula
can be written as: \code{age_formula = ~ I(numeric(obs_time))}, or
another variable for time can be included in auxiliary data.
}
\examples{
sim.dat <- sim_simple_s4t_ch()
m <- fit_ageclass(ageclass ~ FL + obs_time,s4t_ch = sim.dat$s4t_ch)
summary(m)

}
