% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic_loo.R
\name{loo.s4t_cjs_rstan}
\alias{loo.s4t_cjs_rstan}
\title{Efficient approximate leave-one-out cross-validation (LOO)}
\usage{
\method{loo}{s4t_cjs_rstan}(x, ...)
}
\arguments{
\item{x}{a \code{s4t_cjs_rstan} object}

\item{...}{passed to loo::loo}
}
\value{
a named list with class \code{c("psis_loo", "loo")}. See \code{?loo::loo}
}
\description{
Computes the PSIS-LOO CV efficient approximate leave-one-out (LOO)
cross-validation for Bayesian models using Pareto smoothed importance sampling (PSIS).
See documentation for \code{loo::loo()}
}
