% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.clean_s4t_ch}
\alias{print.clean_s4t_ch}
\title{print summary of \code{clean_s4t_ch} object}
\usage{
\method{print}{clean_s4t_ch}(x, ...)
}
\arguments{
\item{x}{\code{clean_s4t_ch} object from \code{clean_s4t_ch()}}

\item{...}{passed to \code{print.tibble()}}
}
\description{
Returns the observations of individuals that had observations dropped
for data cleaning.
}
