% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_summary.R
\name{print.summary.s4t_cjs}
\alias{print.summary.s4t_cjs}
\title{print summary of s4t_cjs}
\usage{
\method{print}{summary.s4t_cjs}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{}
}
\description{
print summary of s4t_cjs
}
