% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_summary.R
\name{summary.s4t_cjs_ml}
\alias{summary.s4t_cjs_ml}
\title{summary of s4t_cjs_ml object}
\usage{
\method{summary}{s4t_cjs_ml}(object, ...)
}
\arguments{
\item{object}{A s4t_cjs_ml object}

\item{...}{Not used}
}
\value{
estimated parameters table and AIC
}
\description{
summary of s4t_cjs_ml object
}
